+++ src/main/java/com/genersoft/iot/vmp/media/service/IMediaServerService.java
@@ -162,7 +162,7 @@ public interface IMediaServerService {
 
     StreamInfo loadMP4File(MediaServer mediaServer, String app, String stream, String datePath);
 
-    void seekRecordStamp(MediaServer mediaServer, String app, String stream, int stamp);
+    void seekRecordStamp(MediaServer mediaServer, String app, String stream, Double stamp);
 
     void setRecordSpeed(MediaServer mediaServer, String app, String stream, Integer speed);
 }
