+++ 数据库/2.7.3/初始化-mysql-2.7.3.sql
@@ -365,7 +365,7 @@ create table IF NOT EXISTS wvp_cloud_record
     file_path       character varying(500),
     collect         bool default false,
     file_size       bigint,
-    time_len        bigint
+    time_len        double precision
 );
 
 drop table IF EXISTS wvp_user;
